refactor: make ucd-generate accessible as a build dep
