+++ src/common/general_category.rs
