+++ src/common/joining_type.rs
