+++ src/common/script.rs
